pyproject.toml
[tool.poetry]
name = "hw27_django"
version = "0.1.0"
description = ""
authors = ["K-Maxim <m.karpenko96@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
